aux-led: blink button LED if temperature is too low/high in standby
In standby mode, blink button LED if ambient temperature is too high or
too low. The comfort zone is set to 18-25C.

Use 6C in off(standby) or lockout mode to toggle if we want to blink
button LED if temperature is out of comfort zone. This configuration is
saved.