fsm-ramping/candle/strobe: do not use dynamic PWM in some strobe modes
Add a flag to temporarily disable dynamic PWM in ramping when we are
in candle mode and firework mode. For candle mode, set the fixed PWM
to 14-bit while using a 12-bit ramping table. This is to allow for a
much smoother light while limiting the max power to 25% to avoid
overheating.