strobe-mode: add a tint alternating strobe mode
For dual channel lights, there is a new mode call tint alternating
mode which switches between channel 1 and 2 periodically with
configurable brightness (the usual 1H and 2H) and interval (4C and 5C
for decreasing/increasing interval by 0.5 sec).

The setting (brightness and interval) is saved and this mode comes
after "tactical strobe".